requirements.txt
flask==3.0.3
requests==2.32.3
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0
